48 Libro Prestamo Completo

+++ routes/web.php
@@ -1,8 +1,5 @@
 <?php
 
-use GuzzleHttp\Middleware;
-use Illuminate\Support\Facades\Route;
-
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -13,13 +10,10 @@ use Illuminate\Support\Facades\Route;
 | contains the "web" middleware group. Now create something great!
 |
 */
+
 /*RUTAS PASSWORD RESET*/
 
-Route::get('password/reset', 'Auth\ForgotPasswordController@showLinkRequestForm')->name('password.request');
-Route::post('password/email', 'Auth\ForgotPasswordController@sendResetLinkEmail')->name('password.email');
-Route::get('password/reset/{token}', 'Auth\ResetPasswordController@showResetForm')->name('password.reset');
-Route::post('password/reset', 'Auth\ResetPasswordController@reset')->name('password.update');
-
+use Illuminate\Support\Facades\Route;
 
 Route::get('password/reset', 'Auth\ForgotPasswordController@showLinkRequestForm')->name('password.request');
 Route::post('password/email', 'Auth\ForgotPasswordController@sendResetLinkEmail')->name('password.email');
@@ -69,18 +63,16 @@ Route::group(['prefix' => 'admin', 'namespace' => 'Admin', 'middleware' => ['aut
     Route::get('permiso-rol', 'PermisoRolController@index')->name('permiso_rol');
     Route::post('permiso-rol', 'PermisoRolController@guardar')->name('guardar_permiso_rol');
 });
-Route::get('libro', 'LibroController@index')->name('libro');
-Route::get('libro/crear', 'LibroController@crear')->name('crear_libro');
-Route::post('libro', 'LibroController@guardar')->name('guardar_libro');
-Route::post('libro/{libro}', 'LibroController@ver')->name('ver_libro');
-Route::get('libro/{id}/editar', 'LibroController@editar')->name('editar_libro');
-Route::put('libro/{id}', 'LibroController@actualizar')->name('actualizar_libro');
-Route::delete('libro/{id}', 'LibroController@eliminar')->name('eliminar_libro');
+Route::get('libro', 'LibroController@index')->name('libro')->middleware('auth');
+Route::get('libro/crear', 'LibroController@crear')->name('crear_libro')->middleware('auth');
+Route::post('libro', 'LibroController@guardar')->name('guardar_libro')->middleware('auth');
+Route::post('libro/{libro}', 'LibroController@ver')->name('ver_libro')->middleware('auth');
+Route::get('libro/{id}/editar', 'LibroController@editar')->name('editar_libro')->middleware('auth');
+Route::put('libro/{id}', 'LibroController@actualizar')->name('actualizar_libro')->middleware('auth');
+Route::delete('libro/{id}', 'LibroController@eliminar')->name('eliminar_libro')->middleware('auth');
 /**
  * Rutas Libro Prestamo
  */
-Route::get('libro-prestamo', 'LibroPrestamoController@index')->name('libro-prestamo');
-
-Auth::routes();
-
-Route::get('/home', 'HomeController@index')->name('home');
+Route::get('libro-prestamo', 'LibroPrestamoController@index')->name('libro-prestamo')->middleware('auth');
+Route::get('libro-prestamo/crear', 'LibroPrestamoController@crear')->name('libro-prestamo.crear')->middleware('auth');
+Route::post('libro-prestamo', 'LibroPrestamoController@guardar')->name('libro-prestamo.guardar')->middleware('auth');
